feat(items): rebuild item detail with sources, recipes, and loot
